refactor(cost-tracker): consolidate PRICING table into _prices.mjs (ADR-149 iter 31 pattern)
Iter 67 added counterfactual.mjs with a comment "single source of truth
would be nicer (iter 31 router pattern); deferred until a third script
needs it." The third script is now cost-projection (iter 66 — uses
session totals, doesn't need per-tier prices yet) and bench.mjs has
its own ANTHROPIC_PRICING for the LLM-baseline path. Two scripts already
share PRICING verbatim; a third write/read will inevitably drift.

This extracts the table NOW, before drift happens.

## Changes

  + scripts/_prices.mjs                   (single source of truth + helpers)
  ~ scripts/track.mjs                     (-15 LOC: import modelTier + costForUsage)
  ~ scripts/counterfactual.mjs            (-15 LOC: import costAtTier)

The underscore-prefix `_prices.mjs` signals "library, not a CLI entry"
— smoke.sh's `scripts/*.mjs` parse check still validates it.

bench.mjs uses a per-test pricing dictionary (`ANTHROPIC_PRICING` with
per-API-version model ids); intentionally left alone for now — when
that drifts from the canonical tier prices, the cost analyst will
catch it via the existing benchmark workflow.

## Verification

Smoke 49/49 still passes (no public surface change).

Math byte-identical to pre-refactor:

  track.mjs synthetic session (10K input + 5K output haiku, 20K input + 10K output sonnet):
    pre/post:  Total cost = \$0.218750  ✓
               (haiku: \$0.0025 + \$0.00625 = \$0.00875; sonnet: \$0.06 + \$0.15 = \$0.21)

  counterfactual.mjs (50K haiku input tokens, \$0.0125 actual):
    pre/post:  always-haiku  \$0.012500  →  +\$0.000000 (0.00%)   ✓
               always-sonnet \$0.150000  →  +\$0.137500 (91.67%)  ✓
               always-opus   \$0.750000  →  +\$0.737500 (98.33%)  ✓

## Why this matters

When pricing rotates (Anthropic price changes, new model tier added),
operators now update exactly ONE file. Iter 31 of the ADR-149 router
work showed that 5 scripts each carrying their own copy meant 4 of
them WOULD eventually be stale on real upgrades. This plugin had 2
scripts with the duplicate; one preempt-extracts the third + makes
the fourth (a future iter) easier.

No behavioral change. No version bump (refactor only — public CLI
surface unchanged).

Source: ADR-149 iter 31 (`v3/@claude-flow/cli/src/ruvector/model-prices.ts`).
