+++ plugins/ruflo-cost-tracker/scripts/track.mjs
@@ -17,33 +17,10 @@ import { readFileSync, writeFileSync, readdirSync, statSync, existsSync } from '
 import { join } from 'node:path';
 import { homedir } from 'node:os';
 import { spawnSync } from 'node:child_process';
+// iter 68 — shared PRICING table (was duplicated in counterfactual.mjs too).
+import { modelTier, costForUsage } from './_prices.mjs';
 
 const PROJECTS_DIR = join(homedir(), '.claude', 'projects');
-
-// USD per 1M tokens — kept in sync with REFERENCE.md "Model pricing" table.
-const PRICING = {
-  haiku:  { input: 0.25,  output: 1.25,  cache_write: 0.30,  cache_read: 0.03 },
-  sonnet: { input: 3.00,  output: 15.00, cache_write: 3.75,  cache_read: 0.30 },
-  opus:   { input: 15.00, output: 75.00, cache_write: 18.75, cache_read: 1.50 },
-};
-
-function modelTier(model) {
-  if (!model) return 'unknown';
-  const m = String(model).toLowerCase();
-  if (m.includes('haiku')) return 'haiku';
-  if (m.includes('sonnet')) return 'sonnet';
-  if (m.includes('opus')) return 'opus';
-  return 'unknown';
-}
-
-function costForUsage(tier, usage) {
-  const p = PRICING[tier];
-  if (!p || !usage) return 0;
-  return (usage.input_tokens || 0) / 1e6 * p.input
-       + (usage.output_tokens || 0) / 1e6 * p.output
-       + (usage.cache_creation_input_tokens || 0) / 1e6 * p.cache_write
-       + (usage.cache_read_input_tokens || 0) / 1e6 * p.cache_read;
-}
 
 function encodeProjectPath(cwd) {
   // #1927: Claude Code encodes an absolute path by replacing path separators
